requirements.txt
viam-sdk
pygame
gtts